Fix bug in export of invoices for accounting
Fixed a bug that prevented to export invoices (accounting). It was
caused by invoices which were generated but belonged to account which
meanwhile disabled invoicing. This flag should be followed only when
generating new invoice. Export has to contain all already generated
invoices.

Method `InvoiceGenerator->renderInvoiceMailAttachment()` refactored.

Generation of invoice file moved to new method which checks if payment
has invoice linked (was generated for payment and probably sent by email
or downloaded by user). This new method `generateInvoiceAsString()` will
be used by export of invoices for accounting and it returns contents of
invoice.

Method `renderInvoiceMailAttachment()` still checks if user wants to have
invoice generated and if invoice isn't ready, generates it and links it
to payment. Returns type of this method doesn't change.

remp/crm#1719