Fix extensive use of file handles when generating invoices ZIP
- PdfResponse always initialized new MPDF instance, which required
lots of files repeatedly. This started to be an issue when generating
export with thousands of invoices. We now pass a single instance
of MPDF to the PDF response so it can reuse it.
- We kept file handle of each tmp PDF file open until it was written
to the ZIP file at the end of the process. This was unnecessary and
we now append PDF to the ZIP right when it's processed and close
the handle to free it up.

remp/crm#1752