Replace company vat ID input with checkbox for CZ invoice forms
Czechs are not familiar with VAT IDs and they use their tax ID
when they need to enter their VAT registration information.

Internal validators previously enforced all tax IDs to be valid in VIES
which was not right, because not all tax ID owners are registered
for VAT.

- We've hidden the input for company VAT ID input if the selected
invoice country is CZ.
- We've added checkbox for CZ to indicate whether the subject is
registered for VAT or not. If it is, we copy the tax ID into the VAT id
field so we don't have to internally change the structure of the fields.

remp/crm#2513