Add internal support for One Stop Shop
- New config added to flag whether the scheme should
be used or not. It's disabled by default, nothing
should be therefore affected.

- If it's enabled, all new payments made via standard
channels (funnel, admin, shop) now try to resolve
the payment country. If it's non-default, the VAT
rates will be adjusted based on the resolved country.
Data providers can be registered by external modules
to add additional checks and resolve logic.

- One stop shop information is now also populated
to the sales funnels via twig variable data provider.
This is primarily to display user information about
the used country.

- In order to get a reasonable default, GeoIP interface
was created with Maxmind implementation as a default.
You need to download and provide path to Maxmind
database in order to use this, or implement your own
implementation of interface.

- The open-source implementation doesn't have the VAT rates
prefilled at the moment. The service to synchronize
them is kept internally for now and will be open-sourced
later. The table vat_rates is created to maintain
compatibility and not break anything.

- Due to the limitations, payments not mathcing payment
country and invoice country are now marked as "not
invoicable".

Full one stop shop support will be added in the near
future, once we test this implementation and resolve
the basic issues.

remp/crm#3191