Fix invalid use of gateways in tests
Lot of tests used fake gateways which were never
correctly registered. This didn't cause any issues
until we added GatewayFactory to the process
of confirmation payment and creation of new subscription.

At that moment, any attempt to confirm the payment
needed to check if the gateway used for payment implements
some interface or not. Since lots of those gateways
weren't registered, tests started to fail.

This fixes the tests and uses universal test gateways
in the affected places.