Move family request VAT data provider to FamilyModule
The original implementation made invoice module dependent
on Family module which was not very OK. We're moving
the data provider to Family module (which is now dependent
on invoice module).

The situation is not ideal as these two modules should
not be dependent on each other, but this direction
is more OK. Ideally, the data provider should only be
registered conditionally if invoices module (or VatModeProvider)
is available.